+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * `rack_response` – prevent response body from yielding `nil`-chunks (@zarqman)
 
+* `derivation_endpoint` – Fix `:upload` option being incompatible with `moving` plugin (@speedo-spin)
+
 ## 2.16.0 (2019-02-18)
 
 * `derivation_endpoint` – Add `:upload_open_options` for download option for derivation result (@janko)